Crypto-Management-System/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0